// src/api/services/product/product.type.ts
export interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  tags: string[];
  brand: string;
  sku: string;
  weight: number;
  dimensions: {
    width: number;
    height: number;
    depth: number;
  };
  warrantyInformation: string;
  shippingInformation: string;
  availabilityStatus: string;
  reviews: {
    rating: number;
    comment: string;
    date: string;
    reviewerName: string;
    reviewerEmail: string;
  }[];
  returnPolicy: string;
  minimumOrderQuantity: number;
  meta: {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
  };
  images: string[];
  thumbnail: string;
}

export interface Category {
  id: number;
  name: string;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  limit: number;
  skip: number;
}

// Define types for query and mutation parameters
export interface GetProductsParams {
  limit: number;
  skip: number;
}

export interface UpdateProductParams {
  id: number;
  data: Partial<Product>;
}